dadude-agent/requirements.txt
# DaDude Agent Dependencies (v2.0 - WebSocket mTLS)

# Web Framework (legacy HTTP, sarà rimosso)
fastapi>=0.104.0
uvicorn>=0.24.0

# WebSocket Client (mTLS)
websockets>=12.0

# WMI (Windows Management)
impacket>=0.11.0

# WinRM (Windows Remote Management)
pywinrm>=0.4.3
requests-ntlm>=1.1.0

# SSH / SFTP
paramiko>=3.3.0

# SNMP
pysnmp>=7.0.0

# MikroTik RouterOS API
routeros_api>=0.17.0

# DNS
dnspython>=2.4.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Security / Encryption
pyjwt>=2.8.0
cryptography>=41.0.0

# Async / Database
aiohttp>=3.9.0
aiosqlite>=0.19.0

# Logging
loguru>=0.7.0

# Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Scheduler
croniter>=2.0.0

# System monitoring
psutil>=5.9.0

